+++ modules/core/builtin/cpp/Gateway.cpp
@@ -25,6 +25,7 @@
 #include "bannerBuiltin.hpp"
 #include "formatBuiltin.hpp"
 #include "execstrBuiltin.hpp"
+#include "evalBuiltin.hpp"
 #include "echoBuiltin.hpp"
 #include "versionBuiltin.hpp"
 #include "narginBuiltin.hpp"
@@ -48,6 +49,7 @@ static const nlsGateway gateway[] =
     { "banner", Nelson::CoreGateway::bannerBuiltin, 0, 0 },
     { "format", Nelson::CoreGateway::formatBuiltin, 1, 1 },
     { "execstr", Nelson::CoreGateway::execstrBuiltin, 1, 2 },
+	{ "eval", Nelson::CoreGateway::evalBuiltin, -1, 2 },
     { "echo", Nelson::CoreGateway::echoBuiltin, 1, 1 },
     { "version", Nelson::CoreGateway::versionBuiltin, 2, 1 },
     { "nargin", Nelson::CoreGateway::narginBuiltin, 1, 1 },