+++ modules/interpreter/src/cpp/Evaluator.cpp
@@ -4264,6 +4264,11 @@ namespace Nelson {
         }
     }
     //=============================================================================
+	bool Evaluator::evaluateString(std::wstring line, bool propogateException)
+	{
+		return evaluateString(wstring_to_utf8(line), propogateException);
+	}
+	//=============================================================================
     bool Evaluator::evaluateString(std::string line, bool propogateException)
     {
         ASTPtr tree = nullptr;